Project Day 2/requirements.txt
PySide6
pynput
pyserial
streamlit
